Create Scala 2 veneer for probe/ProbeValue.scala
