+++ core/src/main/scala/chisel3/probe/ProbeValueBase.scala
@@ -2,15 +2,13 @@
 
 package chisel3.probe
 
-import chisel3.{Data, SourceInfoDoc}
+import chisel3.Data
 import chisel3.internal.Builder
 import chisel3.internal.binding.OpBinding
 import chisel3.internal.firrtl.ir.{ProbeExpr, RWProbeExpr}
 import chisel3.experimental.{requireIsHardware, SourceInfo}
 
-import scala.language.experimental.macros
-
-private[chisel3] sealed trait ProbeValueBase {
+private[chisel3] trait ProbeValueBase {
   protected def apply[T <: Data](source: T, writable: Boolean)(implicit sourceInfo: SourceInfo): T = {
     requireIsHardware(source)
     // construct probe to return with cloned info
@@ -32,22 +30,4 @@ private[chisel3] sealed trait ProbeValueBase {
     }
     clone
   }
-}
-
-object ProbeValue extends ProbeValueBase with SourceInfoDoc {
-
-  /** Create a read-only probe expression. */
-  def apply[T <: Data](source: T): T = macro chisel3.internal.sourceinfo.ProbeTransform.sourceApply[T]
-
-  /** @group SourceInfoTransformMacro */
-  def do_apply[T <: Data](source: T)(implicit sourceInfo: SourceInfo): T = super.apply(source, writable = false)
-}
-
-object RWProbeValue extends ProbeValueBase with SourceInfoDoc {
-
-  /** Create a read/write probe expression. */
-  def apply[T <: Data](source: T): T = macro chisel3.internal.sourceinfo.ProbeTransform.sourceApply[T]
-
-  /** @group SourceInfoTransformMacro */
-  def do_apply[T <: Data](source: T)(implicit sourceInfo: SourceInfo): T = super.apply(source, writable = true)
 }
